requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
WMI>=1.5.1
pyinstaller>=6.0.0
